only configure fontconfig if the etc/fonts/font.conf file was deployed

+++ AppImageBuilder/drivers/FontConfig.py
@@ -9,6 +9,7 @@
 #
 #  The above copyright notice and this permission notice shall be included in
 #  all copies or substantial portions of the Software.
+import os
 
 from AppImageBuilder import drivers
 
@@ -17,20 +18,22 @@ class FontConfig(drivers.Driver):
     id = 'fontconfig'
 
     def configure(self, app_dir):
-        app_dir.app_run.env['FONTCONFIG_FILE'] = '${APPDIR}/etc/fonts/fonts.conf'
-        app_dir.app_run.env['FONTCONFIG_PATH'] = '${APPDIR}/usr/share/fontconfig'
-        app_dir.app_run.env['FONTCONFIG_SYSROOT'] = '${APPDIR}'
+        fonts_conf_path = app_dir.path + '/etc/fonts/fonts.conf'
+        if os.path.exists(fonts_conf_path):
+            app_dir.app_run.env['FONTCONFIG_FILE'] = '${APPDIR}/etc/fonts/fonts.conf'
+            app_dir.app_run.env['FONTCONFIG_PATH'] = '${APPDIR}/usr/share/fontconfig'
+            app_dir.app_run.env['FONTCONFIG_SYSROOT'] = '${APPDIR}'
 
-        lines = []
-        with open(app_dir.path + '/etc/fonts/fonts.conf', 'r') as f:
-            lines.extend(f.readlines())
+            lines = []
+            with open(fonts_conf_path, 'r') as f:
+                lines.extend(f.readlines())
 
-        new_lines = []
-        for line in lines:
-            new_lines.append(line)
+            new_lines = []
+            for line in lines:
+                new_lines.append(line)
 
-            if '<!-- Font directory list -->' in line:
-                new_lines.append('<dir prefix="APPDIR">usr/share/fonts</dir>')
+                if '<!-- Font directory list -->' in line:
+                    new_lines.append('<dir prefix="APPDIR">usr/share/fonts</dir>')
 
-        with open(app_dir.path + '/etc/fonts/fonts.conf', 'w') as f:
-            f.writelines(new_lines)
+            with open(app_dir.path + '/etc/fonts/fonts.conf', 'w') as f:
+                f.writelines(new_lines)
